+++ sw.js
@@ -1,8 +1,4 @@
-/*
-    the PWA service worker is currently not used because of a memory leak
-    issue introduced in Chrome v129
-*/
-var snapVersion = '10.0.11', // currently unused
+var snapVersion = '10.0.11',
     cacheName = `snap-pwa-${snapVersion}`,
     filesToCache = [
         'snap.html',